refactor: Split `ìnner` module into `accounts` and `deal_proposals`, and rename `inner` to `benchmarking`
