+++ pallets/storage-provider/benchmarks/src/accounts.rs
@@ -1,0 +1,14 @@
+use sp_io::crypto::ed25519_generate;
+use sp_runtime::{traits::IdentifyAccount, AccountId32, MultiSigner};
+
+pub const ALICE: &'static str = "//Alice";
+pub const STORAGE_PROVIDER: &'static str = "//StorageProvider";
+
+pub fn generate_benchmark_account(name: &'static str) -> (AccountId32, MultiSigner) {
+    let signer: sp_core::ed25519::Public =
+        ed25519_generate(0.into(), Some(name.as_bytes().to_vec())).into();
+    let signer: MultiSigner = signer.into();
+    let account_id = signer.clone().into_account();
+
+    (account_id, signer)
+}
