+++ pallets/storage-provider/benchmarks/src/lib.rs
@@ -2,14 +2,17 @@
 
 extern crate alloc;
 
+#[cfg(any(feature = "runtime-benchmarks", test))]
+mod accounts;
+
 #[cfg(feature = "runtime-benchmarks")]
-pub mod inner;
+pub mod benchmarking;
+
+#[cfg(any(feature = "runtime-benchmarks", test))]
+mod deal_proposals;
 
 #[cfg(all(feature = "runtime-benchmarks", test))]
 pub(crate) mod mock;
-
-#[cfg(any(feature = "runtime-benchmarks", test))]
-mod utils;
 
 #[cfg(any(feature = "runtime-benchmarks", test))]
 mod pallet;
